@@ -1,12 +1,9 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createRouter, createWebHashHistory } from 'vue-router'
 import Flags from '@/components/Flags'
 import Flag from '@/components/Flag'
 
-Vue.use(Router)
-
-export default new Router({
-  mode: 'hash',
+export default createRouter({
+  history: createWebHashHistory(),
   routes: [
     {
       path: '/',
